moved calculator into is proper folder
