+++ src-back/Service/Calculator/Calculator.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service;
+namespace App\Service\Calculator;
 
 use ParseError;
