+++ src-back/Service/Calculator/CalculatorException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service;
+namespace App\Service\Calculator;
 
 class CalculatorException extends \Exception
 {
