+++ src-back/Service/Calculator/CalculatorInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service;
+namespace App\Service\Calculator;
 
 interface CalculatorInterface
 {
